chore: refactor json utils
