+++ utils/index.js
@@ -1,6 +1,5 @@
 'use strict';
 
-// const resolvePath = require('./resolve-path');
 const sendCommand = require('./send-command');
 const logger = require('./logger');
 const calculateFileHash = require('./calculate-file-hash');
@@ -9,15 +8,11 @@ const compareHashes = require('./compare-hashes');
 const fetchLatestVersion = require('./fetch-latest-version');
 const fetchPackageMeta = require('./fetch-package-meta');
 const incrementSemverVersion = require('./increment-semver-version');
-const readJSON = require('./read-json');
-const writeJSON = require('./write-json');
 const compressedSize = require('./compressed-size');
 const getDefaults = require('./get-defaults');
 const getCWD = require('./get-cwd');
-const writeEikJSON = require('./write-eik-json');
 
 module.exports = {
-    // resolvePath,
     sendCommand,
     logger,
     calculateFileHash,
@@ -26,10 +21,7 @@ module.exports = {
     fetchLatestVersion,
     fetchPackageMeta,
     incrementSemverVersion,
-    writeJSON,
-    readJSON,
     compressedSize,
     getDefaults,
     getCWD,
-    writeEikJSON,
 };
