+++ utils/json/index.js
@@ -1,0 +1,5 @@
+const read = require('./read');
+const write = require('./write');
+const writeEik = require('./write-eik');
+
+module.exports = { read, write, writeEik };
