+++ utils/json/write-eik.js
